backend/requirements.txt
fastapi>=0.103.2
uvicorn>=0.15.0
pymongo>=3.12.0
pydantic>=2.4.1
python-jose>=3.3.0
PyJWT>=2.1.0
pymongo[srv]>=3.12.0
